+++ src/Routing/FastRouteDispatcher.php
@@ -2,6 +2,10 @@
 
 namespace Rougin\Slytherin\Routing;
 
+use FastRoute\Dispatcher\GroupCountBased;
+use FastRoute\RouteCollector;
+use FastRoute\RouteParser\Std;
+
 /**
  * FastRoute Dispatcher
  *
@@ -12,7 +16,7 @@ namespace Rougin\Slytherin\Routing;
  * @package Slytherin
  * @author  Rougin Royce Gutib <rougingutib@gmail.com>
  */
-class FastRouteDispatcher implements DispatcherInterface
+class FastRouteDispatcher extends AbstractDispatcher implements DispatcherInterface
 {
     /**
      * @var \FastRoute\Dispatcher
@@ -23,16 +27,6 @@ class FastRouteDispatcher implements DispatcherInterface
      * @var \Rougin\Slytherin\Routing\RouterInterface
      */
     protected $router;
-
-    /**
-     * Initializes the dispatcher instance.
-     *
-     * @param \Rougin\Slytherin\Routing\RouterInterface|null $router
-     */
-    public function __construct(RouterInterface $router = null)
-    {
-        $router == null || $this->router($router);
-    }
 
     /**
      * Dispatches against the provided HTTP method verb and URI.
@@ -72,29 +66,12 @@ class FastRouteDispatcher implements DispatcherInterface
 
         $this->router = $router;
 
-        $this->dispatcher = \FastRoute\simpleDispatcher($routes);
+        $dispatcher = new \FastRoute\DataGenerator\GroupCountBased;
+
+        $routes($router = new RouteCollector(new Std, $dispatcher));
+
+        $this->dispatcher = new GroupCountBased($router->getData());
 
         return $this;
-    }
-
-    /**
-     * Checks if the specified method is a valid HTTP method.
-     *
-     * @param  string $httpMethod
-     * @return boolean
-     *
-     * @throws UnexpectedValueException
-     */
-    protected function allowed($httpMethod)
-    {
-        $allowed = array('DELETE', 'GET', 'OPTIONS', 'PATCH', 'POST', 'PUT');
-
-        if (in_array($httpMethod, $allowed) === false) {
-            $message = 'Used method is not allowed';
-
-            throw new \UnexpectedValueException($message);
-        }
-
-        return true;
     }
 }